+++ Loop/pass/analysispass.cpp
@@ -25,6 +25,9 @@
 #include <unordered_map>
 #include <utility>
 
+#include "features.hpp"
+
+
 using std::unordered_map;
 using std::vector;
 using std::pair;
@@ -39,11 +42,12 @@ namespace {
             _loop = li.getLoopFor(i->getParent());
             _I = i;
 
-            initialize(li);
-            //errs() << "Instruction " << i << " is with loop " << (_loop ? std::to_string(reinterpret_cast<size_t>(_loop)) : " No loops") << "\n";
+            initialize_loopanalysis(li);
+
+            // Add Data Analysis and Control Analysis here
         }
 
-        void initialize(llvm::LoopAnalysis::Result& _li) {
+        void initialize_loopanalysis(llvm::LoopAnalysis::Result& _li) {
             if(!_loop){
                 errs() << "This branch instruction does not have a loop associated with it.\n";
                 return;
@@ -117,43 +121,7 @@ namespace {
             errs() << "\n" << _loopinfo << "\n\n";
         }
 
-        struct LoopFeatures {
-            LoopFeatures()  {}
-
-            // Numerical Features
-            int loop_depth;
-            int number_BB;
-            int number_exits; // These are exit edges
-            int number_exit_blocks; // These are exit blocks
-
-            // Add Features Not in Paper
-            int num_successors;
-            
-
-            // Categorical Features
-            bool isexit; 
-            bool isbackedge;
-            bool isdestinationinloop;
-            bool isdestinationnestedloop;
-
-            friend llvm::raw_ostream &operator << (llvm::raw_ostream &os, const LoopFeatures &lf){
-
-                os << "Loop Depth: " << lf.loop_depth << "\n";
-                os << "Number of BBs: " << lf.number_BB << "\n";
-                os << "Number of Exits: " << lf.number_exits << "\n";
-                os << "NUmber of Exit Blocks: " << lf.number_exit_blocks << "\n";
-                os << "NUmber of Successors: " << lf.num_successors << "\n";
-
-
-                os << "Is Exit?: " << (lf.isexit? " Yes " : " No ") << "\n";
-                os << "Is Back Edge?: " << (lf.isbackedge? " Yes " : " No ") << "\n";
-                os << "Is the Destination In Loop?: " << (lf.isdestinationinloop? " Yes " : " No ") << "\n";
-                os << "Is the Destination In Nested Loop?: " << (lf.isdestinationnestedloop? " Yes " : " No ") << "\n";
-
-                return os;
-            }
-
-        };
+        
 
         LoopFeatures _loopinfo;
         llvm::Loop* _loop;
@@ -182,12 +150,6 @@ namespace {
                 }
             }
             return PreservedAnalyses::all();
-        }
-
-
-        void write() {
-            //TODO: write to CSV? or what should we do? format needs to be decided :)
-            return;
         }
 
     };
