+++ Loop/tests/fiveloops.c
@@ -1,0 +1,23 @@
+#include <stdio.h>
+
+int main() {
+    for(int i=0; i<10; i++){
+
+        for(int j=0; j<10; j++){
+
+            for(int k=0; k<12; k++){
+
+                for(int z=0;z>5; z++){
+
+                    for(int u=1; u<=9; u++){
+
+                        printf("I don't know");
+
+                    }
+                }
+            }
+        }
+    }
+
+    return 0;
+}
