+++ Loop/todo.md
@@ -1,0 +1,5 @@
+
+- [ ] Encode Opcodes into Numerical Values
+- [ ] Encode Variables into Special Values
+- [ ] Merge Branch - Loop & Data
+- [ ] Decide Output Data Format
